refactor(rename): spelling reference correctly
