+++ lib/AdvancementsParser.js
@@ -1,7 +1,7 @@
 const fs = require("fs");
 const path = require("path");
 
-const PATHS = require("./helpers/PathRefrence").paths;
+const PATHS = require("./helpers/PathReference").paths;
 const logger = require("./helpers/Logger").getLogger();
 
 const merge = require("deepmerge");
