+++ lib/Configuration.js
@@ -72,7 +72,7 @@ logger.debug(`Current working dir ${RUNNING_DIR}`, { domain: DOMAIN });
 logger.silly(`Provided CLI arguments:\n${JSON.stringify(ARGV, null, "  ")}`, { domain: DOMAIN });
 
 // SETUP PATH REFERENCES
-const pathref = require("./helpers/PathRefrence");
+const pathref = require("./helpers/PathReference");
 pathref.setBasePaths(ARGV.minecraftdir, ARGV.outputdir, ARGV.workdir);
 const PATHS = pathref.paths;
