+++ lib/JarDataExtractor.js
@@ -1,7 +1,7 @@
 const Zip = require("node-7z");
 const ZipBin = require("7zip-bin");
 
-const PATHS = require("./helpers/PathRefrence").paths;
+const PATHS = require("./helpers/PathReference").paths;
 const logger = require("./helpers/Logger").getLogger();
 const DOMAIN = "JarDataExtractor";
