+++ lib/LogsParser.js
@@ -1,4 +1,4 @@
-const PATHS = require("./helpers/PathRefrence").paths;
+const PATHS = require("./helpers/PathReference").paths;
 const logger = require("./helpers/Logger").getLogger();
 const logconst = require("./helpers/LogConst");
 const logregx = require("./helpers/LogsRegex");
