+++ lib/McaParser.js
@@ -2,7 +2,7 @@ const path = require("path");
 const fs = require("fs");
 
 const NbtTools = require("./NbtTools");
-const PATHS = require("./helpers/PathRefrence").paths;
+const PATHS = require("./helpers/PathReference").paths;
 const logger = require("./helpers/Logger").getLogger();
 
 const mca = require("mca-js");
