+++ lib/PlayerData.js
@@ -2,7 +2,7 @@ const fs = require("fs");
 const path = require("path");
 
 const NbtTools = require("./NbtTools");
-const PATHS = require("./helpers/PathRefrence").paths;
+const PATHS = require("./helpers/PathReference").paths;
 const logger = require("./helpers/Logger").getLogger();
 
 const DOMAIN = "PlayerData";
