+++ lib/ProfileHelper.js
@@ -1,5 +1,5 @@
 const { PLAYERS, ACCEPTABLE_PROFILE_AGE } = require("./Configuration");
-const PATHS = require("./helpers/PathRefrence").paths;
+const PATHS = require("./helpers/PathReference").paths;
 const logger = require("./helpers/Logger").getLogger();
 const MojangAPI = require("./MojangApi");
 const path = require("path");
