+++ lib/ServerDataTool.js
@@ -5,7 +5,7 @@
 const fs = require("fs");
 const path = require("path");
 const JarExtractor = require("./JarDataExtractor");
-const PATHS = require("./helpers/PathRefrence").paths;
+const PATHS = require("./helpers/PathReference").paths;
 const logger = require("./helpers/Logger").getLogger();
 const NbtTools = require("./NbtTools");
