+++ lib/helpers/PathReference.js
@@ -69,7 +69,7 @@ function resolvePaths() {
   RESOLVED_PATHS.OPSLIST_FILE = path.join(RESOLVED_PATHS.MC_DIR, "ops.json");
   RESOLVED_PATHS.LOGS_DIR = path.join(RESOLVED_PATHS.MC_DIR, "logs");
   RESOLVED_PATHS.TEMP_DIR = path.join(RESOLVED_PATHS.WORK_DIR, ".temp");
-  RESOLVED_PATHS.EXTRACTED_DIR = path.join(BASE_PATHS.WORK_DIR, "extracted");
+  RESOLVED_PATHS.EXTRACTED_DIR = path.join(RESOLVED_PATHS.WORK_DIR, "extracted");
   RESOLVED_PATHS.EXTRACTED_ASSETS_DIR = path.join(RESOLVED_PATHS.EXTRACTED_DIR, "assets");
   RESOLVED_PATHS.EXTRACTED_DATA_DIR = path.join(RESOLVED_PATHS.EXTRACTED_DIR, "data");
   RESOLVED_PATHS.EXTRACTED_DATA_MINECRAFT_DIR = path.join(RESOLVED_PATHS.EXTRACTED_DATA_DIR, "minecraft");
